requirements.txt
pillow>=8.0.0
pygame>=2.0.0